Changed comments and added small animation
When you click on post you no longer get grayed out, you stay in the same place. All the old code is still there so you can easily change it back.

Also addad an animation to the bars so that you get a bit of responce on your new score.